# openapi-mvp.yaml
openapi: 3.0.3
info:
  title: Puyuan AI Platform API (MVP)
  version: 1.0.0
  description: MVP API for merchant platform, billing, plugins and admin operations.
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login by mobile code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
  /api/v1/tenant/profile:
    get:
      tags: [Auth]
      summary: Get current tenant and user profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileResponse'
...
  /api/v1/billing/statements/daily:
    get:
      tags: [Billing]
      summary: Query daily statement
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XRequestId'
        - in: query
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Daily statement
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DailyStatement'
  /api/v1/billing/statements/monthly:
    get:
      tags: [Billing]
      summary: Query monthly statement
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XRequestId'
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^[0-9]{4}-(0[1-9]|1[0-2])$'
      responses:
        '200':
          description: Monthly statement
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MonthlyStatement'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XTenantId:
      in: header
      name: X-Tenant-Id
      required: true
      schema:
        type: string
    XRequestId:
      in: header
      name: X-Request-Id
      required: true
      schema:
        type: string
...
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          type: object
          nullable: true
        request_id:
          type: string
      required: [code, message, request_id]
    LoginRequest:
      type: object
      properties:
        mobile:
          type: string
        verify_code:
          type: string
      required: [mobile, verify_code]
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
        user_id:
          type: integer
          format: int64
        tenant_id:
          type: integer
          format: int64
        role_code:
          type: string
    ProfileResponse:
      type: object
      properties:
        tenant_id:
          type: integer
          format: int64
        tenant_code:
          type: string
        tenant_name:
          type: string
        tenant_status:
          type: integer
        user_id:
          type: integer
          format: int64
        role_code:
          type: string
...
    DailyStatement:
      type: object
      properties:
        stat_date:
          type: string
          format: date
        token_in:
          type: integer
        token_out:
          type: integer
        call_count:
          type: integer
        amount_recharge:
          type: number
          format: double
        amount_refund:
          type: number
          format: double
    MonthlyStatement:
      type: object
      properties:
        month:
          type: string
        token_in:
          type: integer
        token_out:
          type: integer
        call_count:
          type: integer
        amount_recharge:
          type: number
          format: double
        amount_refund:
          type: number
          format: double
...